backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
httpx==0.19.0
web3==5.24.0
pydantic==1.8.2
python-dateutil==2.8.2
blockcypher==1.0.93